+++ turbo/apps/platform/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [0.165.1](https://github.com/vm0-ai/vm0/compare/app-v0.165.0...app-v0.165.1) (2026-03-30)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @vm0/core bumped to 8.126.0
+
 ## [0.165.0](https://github.com/vm0-ai/vm0/compare/app-v0.164.0...app-v0.165.0) (2026-03-30)